dataframe: move io/optimize out of core
