+++ dask/dataframe/__init__.py
@@ -1,4 +1,4 @@
-from .core import (DataFrame, Series, Index, _Frame, from_array,
-        from_bcolz, concat)
+from .core import (DataFrame, Series, Index, _Frame, concat)
 from .shuffle import from_pframe
-from .io import read_csv
+from .io import (read_csv, from_array, from_bcolz, from_array,
+        from_bcolz)
